+++ pyfakefs/mox3_stubout.py
@@ -102,7 +102,7 @@ class StubOutForTesting:
         # function. We need to ensure that we put it back as a staticmethod.
         old_attribute = obj.__dict__.get(attr_name)
         if old_attribute is not None and isinstance(old_attribute, staticmethod):
-            orig_attr = staticmethod(orig_attr)
+            orig_attr = staticmethod(orig_attr)  # pytype: disable=not-callable
 
         self.stubs.append((orig_obj, attr_name, orig_attr))
         setattr(orig_obj, attr_name, new_attr)