+++ pyfakefs/fake_filesystem_unittest.py
@@ -36,6 +36,7 @@ pyfakefs by simply changing their base class from `:py:class`unittest.TestCase`
 to `:py:class`pyfakefs.fake_filesystem_unittest.TestCase`.
 """
 import _io  # type:ignore[import]
+import builtins
 import doctest
 import functools
 import genericpath
@@ -455,7 +456,7 @@ class Patcher:
         os.path,
     }
     if sys.platform == "win32":
-        import nt  # type:ignore [import]
+        import nt  # type:ignore[import]
         import ntpath
 
         SKIPMODULES.add(nt)
@@ -559,6 +560,7 @@ class Patcher:
         # save the original open function for use in pytest plugin
         self.original_open = open
         self.patch_open_code = patch_open_code
+        self.fake_open: fake_open.FakeFileOpen
 
         if additional_skip_names is not None:
             skip_names = [
@@ -844,6 +846,7 @@ class Patcher:
 
         self.fs = fake_filesystem.FakeFilesystem(patcher=self, create_temp_dir=True)
         self.fs.patch_open_code = self.patch_open_code
+        self.fake_open = fake_open.FakeFileOpen(self.fs)
         for name in self._fake_module_classes:
             self.fake_modules[name] = self._fake_module_classes[name](self.fs)
             if hasattr(self.fake_modules[name], "skip_names"):
@@ -922,6 +925,9 @@ class Patcher:
             for module, attr in modules:
                 if attr in self.unfaked_modules:
                     self._stubs.smart_set(module, name, self.unfaked_modules[attr])
+        if sys.version_info >= (3, 12):
+            # workaround for patching open - does not work with skip modules
+            self._stubs.smart_set(builtins, "open", self.fake_open)
 
     def patch_defaults(self) -> None:
         for fct, idx, ft in self.FS_DEFARGS: