+++ pyfakefs/tests/fake_filesystem_unittest_test.py
@@ -28,7 +28,6 @@ import sys
 import tempfile
 import unittest
 import warnings
-from distutils.dir_util import copy_tree, remove_tree
 from pathlib import Path
 from unittest import TestCase, mock
 
@@ -43,6 +42,10 @@ from pyfakefs.fake_filesystem_unittest import (
     PatchMode,
 )
 from pyfakefs.tests.fixtures import module_with_attributes
+
+if sys.version_info < (3, 12):
+    # distutils removed in Python 3.12
+    from distutils.dir_util import copy_tree, remove_tree
 
 
 class TestPatcher(TestCase):
@@ -152,9 +155,9 @@ class TestPyfakefsUnittest(TestPyfakefsUnittestBase):  # pylint: disable=R0904
         self.assertFalse(self.fs.exists("/test/dir1"))
 
     def test_fakepathlib(self):
-        with pathlib.Path("/fake_file.txt") as p:
-            with p.open("w") as f:
-                f.write("text")
+        p = pathlib.Path("/fake_file.txt")
+        with p.open("w") as f:
+            f.write("text")
         is_windows = sys.platform.startswith("win")
         if is_windows:
             self.assertTrue(self.fs.exists(r"\fake_file.txt"))
@@ -225,12 +228,14 @@ class TestPatchingImports(TestPyfakefsUnittestBase):
         stat_result = pyfakefs.tests.import_as_example.file_stat2(file_path)
         self.assertEqual(3, stat_result.st_size)
 
+    @unittest.skipIf(sys.version_info >= (3, 12), "Currently not working in 3.12")
     def test_import_open_as_other_name(self):
         file_path = "/foo/bar"
         self.fs.create_file(file_path, contents=b"abc")
         contents = pyfakefs.tests.import_as_example.file_contents1(file_path)
         self.assertEqual("abc", contents)
 
+    @unittest.skipIf(sys.version_info >= (3, 12), "Currently not working in 3.12")
     def test_import_io_open_as_other_name(self):
         file_path = "/foo/bar"
         self.fs.create_file(file_path, contents=b"abc")
@@ -393,6 +398,10 @@ class AdditionalSkipNamesTest(fake_filesystem_unittest.TestCase):
         self.fs.create_file("foo")
         self.assertFalse(pyfakefs.tests.import_as_example.check_if_exists7("foo"))
 
+    @unittest.skipIf(
+        sys.version_info >= (3, 12),
+        "Skip modules currently not working for open in 3.12",
+    )
     def test_open_succeeds(self):
         pyfakefs.tests.import_as_example.open_this_file()
 
@@ -438,6 +447,10 @@ class AdditionalSkipNamesModuleTest(fake_filesystem_unittest.TestCase):
         self.fs.create_file("foo")
         self.assertFalse(pyfakefs.tests.import_as_example.check_if_exists7("foo"))
 
+    @unittest.skipIf(
+        sys.version_info >= (3, 12),
+        "Skip modules currently not working for open in 3.12",
+    )
     def test_open_succeeds(self):
         pyfakefs.tests.import_as_example.open_this_file()
 
@@ -704,31 +717,33 @@ class TestShutilWithZipfile(fake_filesystem_unittest.TestCase):
         shutil.make_archive("archive", "zip", root_dir="foo")
 
 
-class TestDistutilsCopyTree(fake_filesystem_unittest.TestCase):
-    """Regression test for #501."""
+if sys.version_info < (3, 12):
 
-    def setUp(self):
-        self.setUpPyfakefs()
-        self.fs.create_dir("./test/subdir/")
-        self.fs.create_dir("./test/subdir2/")
-        self.fs.create_file("./test2/subdir/1.txt")
+    class TestDistutilsCopyTree(fake_filesystem_unittest.TestCase):
+        """Regression test for #501."""
 
-    def test_file_copied(self):
-        copy_tree("./test2/", "./test/")
-        remove_tree("./test2/")
+        def setUp(self):
+            self.setUpPyfakefs()
+            self.fs.create_dir("./test/subdir/")
+            self.fs.create_dir("./test/subdir2/")
+            self.fs.create_file("./test2/subdir/1.txt")
 
-        self.assertTrue(os.path.isfile("./test/subdir/1.txt"))
-        self.assertFalse(os.path.isdir("./test2/"))
+        def test_file_copied(self):
+            copy_tree("./test2/", "./test/")
+            remove_tree("./test2/")
 
-    def test_file_copied_again(self):
-        # used to fail because 'test2' could not be found
-        self.assertTrue(os.path.isfile("./test2/subdir/1.txt"))
+            self.assertTrue(os.path.isfile("./test/subdir/1.txt"))
+            self.assertFalse(os.path.isdir("./test2/"))
 
-        copy_tree("./test2/", "./test/")
-        remove_tree("./test2/")
+        def test_file_copied_again(self):
+            # used to fail because 'test2' could not be found
+            self.assertTrue(os.path.isfile("./test2/subdir/1.txt"))
 
-        self.assertTrue(os.path.isfile("./test/subdir/1.txt"))
-        self.assertFalse(os.path.isdir("./test2/"))
+            copy_tree("./test2/", "./test/")
+            remove_tree("./test2/")
+
+            self.assertTrue(os.path.isfile("./test/subdir/1.txt"))
+            self.assertFalse(os.path.isdir("./test2/"))
 
 
 class PathlibTest(TestCase):