+++ CHANGES.md
@@ -12,6 +12,10 @@ The released versions correspond to PyPI releases.
   see [#830](../../issues/830) and [#831](../../issues/831)).
 * Adapt to changes in `shutil` in Python 3.12 beta2 (see [#814](../../issues/814)).
 
+### Documentation
+* Added a note regarding the incompatibility of the built-in `sqlite3` module with
+  `pyfakefs` (see [#850](../../issues/850))
+
 ### Infrastructure
 * Added pytype check for non-test modules in CI (see [#599](../../issues/599)).